requirements.txt
streamlit
agno
python-dotenv
duckduckgo-search
lancedb
tantivy
pypdf
sqlalchemy
yfinance
groq
openai